+++ src/qbits/alia/codec.clj
@@ -72,17 +72,18 @@ not (yet) supports them. See
 https://issues.apache.org/jira/browse/CASSANDRA-3783"))))
 
 (defn result-set->maps
-  [^ResultSet result-set]
-  (-> (map (fn [^Row row]
-             (let [cdef (.getColumnDefinitions row)
-                   len (.size cdef)]
-               (loop [idx (int 0)
-                      row-map (transient {})]
-                 (if (= idx len)
-                   (persistent! row-map)
-                   (recur (int (inc idx))
-                          (assoc! row-map
-                            (.getName cdef idx)
-                            (decode row idx (.getType cdef idx))))))))
-            result-set)
-      (vary-meta assoc :query-trace (.getQueryTrace result-set))))
+  [^ResultSet result-set keywordize?]
+  (let [key-mod  (if keywordize? keyword identity)]
+    (-> (map (fn [^Row row]
+               (let [cdef (.getColumnDefinitions row)
+                     len (.size cdef)]
+                 (loop [idx (int 0)
+                        row-map (transient {})]
+                   (if (= idx len)
+                     (persistent! row-map)
+                     (recur (int (inc idx))
+                            (assoc! row-map
+                                    (key-mod (.getName cdef idx))
+                                    (decode row idx (.getType cdef idx))))))))
+             result-set)
+        (vary-meta assoc :query-trace (.getQueryTrace result-set)))))